extract basic coroutine usecase to core module
