+++ modules/core/src/main/java/lu/aqu/core/coroutine/CoroutineUseCase.kt
@@ -1,0 +1,11 @@
+package lu.aqu.core.coroutine
+
+interface CoroutineUseCase<T> {
+
+    fun execute(
+        onLoading: () -> Unit = {},
+        onResult: (T) -> Unit,
+        onError: (Throwable) -> Unit = {},
+        onFinished: () -> Unit = {}
+    )
+}
