+++ modules/features/project/usecase/src/main/java/lu/aqu/projper/project/usecase/FindProjectsUseCase.kt
@@ -1,13 +1,6 @@
 package lu.aqu.projper.project.usecase
 
+import lu.aqu.core.coroutine.CoroutineUseCase
 import lu.aqu.projper.project.domain.Project
 
-interface FindProjectsUseCase {
-
-    fun execute(
-        onLoading: () -> Unit = {},
-        onResult: (List<Project>) -> Unit,
-        onError: (Throwable) -> Unit = {},
-        onFinished: () -> Unit = {}
-    )
-}
+interface FindProjectsUseCase : CoroutineUseCase<List<Project>>
