+++ modules/features/project/usecase/src/main/java/lu/aqu/projper/project/usecase/impl/FindProjectsUseCaseImpl.kt
@@ -1,9 +1,7 @@
 package lu.aqu.projper.project.usecase.impl
 
 import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
+import lu.aqu.core.coroutine.CoroutineUseCaseAbs
 import lu.aqu.projper.project.domain.Project
 import lu.aqu.projper.project.domain.ProjectRepository
 import lu.aqu.projper.project.usecase.FindProjectsUseCase
@@ -11,29 +9,10 @@ import javax.inject.Inject
 
 internal class FindProjectsUseCaseImpl @Inject constructor(
     private val projectRepository: ProjectRepository,
-    private val mainDispatcher: CoroutineDispatcher,
-    private val ioDispatcher: CoroutineDispatcher
-) : FindProjectsUseCase {
+    mainDispatcher: CoroutineDispatcher,
+    ioDispatcher: CoroutineDispatcher
+) : CoroutineUseCaseAbs<List<Project>>(mainDispatcher, ioDispatcher), FindProjectsUseCase {
 
-    override fun execute(
-        onLoading: () -> Unit,
-        onResult: (List<Project>) -> Unit,
-        onError: (Throwable) -> Unit,
-        onFinished: () -> Unit
-    ) {
-        CoroutineScope(mainDispatcher).launch {
-            onLoading()
-
-            try {
-                val result = withContext(ioDispatcher) {
-                    projectRepository.findAll()
-                }
-                onResult(result)
-            } catch (exception: Exception) {
-                onError(exception)
-            }
-
-            onFinished()
-        }
-    }
+    override suspend fun executeAsync(): List<Project> =
+        projectRepository.findAll()
 }
